Improved docker files to allow for a better work flow
Work flow should now be to run ./run.sh. This file now has a version
tag inside it. If any of the Dockerfile's or the run.sh file itself
is changed in a way that affects the devs and requires them to rebuild
their docker images/virtual machines, you can just add 1 to the version
number in the file. This will cause the next time the run.sh file to be
ran to look for anything images matching the new name which will fail
and remove any old versions of the image, then rebuild from scratch.

This patch also will auto run the "make" command when run.sh is ran.

There is also a new Dockerfile in DockerBase directory. This is to
allow docker to build a base image with the build dependencies and use
it's cached version for each itteration run. This should dramatically
improve speed to build.

+++ run.sh
@@ -1,7 +1,17 @@
-docker build --tag trader .;
+#!/bin/sh
+version="1.0";
 
-docker rm trader;
+if [[ -z $(docker images -a | grep "trader_base_$version ") ]] ; then
+    docker images -a | grep trader_ | awk '{print $1}' | xargs docker rmi
+    docker build --tag "trader_base_$version" DockerBase;
 
-docker create --name trader trader;
+    data=$(sed "1s/.*/FROM trader_base_$version/" ./Dockerfile)
+    echo "${data}" > ./Dockerfile;
 
-docker run -i -t trader bash
+    docker rmi -f trader;
+    docker build --tag trader .;
+    docker rm trader;
+    docker create --name trader trader;
+fi
+
+docker run -v `pwd`/:/trader -i -t trader
